undo/redo improvements
migrated the entire undo/redo system to use clean state! this means we can now carry over the undo stack when saving the map and it makes handling redos a bit cleaner. i also implemented redos for tile placement - no events yet, but i'll get around to it! the event data setting code has some oddities that make this another task that requires a small rewrite. i swear every single piece of code in this editor was written to fit one extremely niche use case and break on everything else...

as for the tile placement undo/redo behavior: right now it stores a full copy of the layer's tiles before and after the operation, which is pretty wasteful - perhaps a migration to a hashmap is in order, where the key would be the tile index in the map and the value would be the tile id.

this commit also fixes some jank with the shift mode by detecting the shift key press the right way.