terrain widget work continued
the lcf widget bindings are a blessing. this is one of the reasons why i'm making this project out of the existing editor and pushing through the rewrites instead of spinning up my own project.

i'm not too happy about the current solution for selecting music. i think this should be handled by a dedicated widget that behaves like the one in the vanilla editor which holds a Sound instance and can be setup with one connect/bind pair. i'll work on that later and just skip adding the footstep sound picker for now...

not exactly looking forward to writing a grid renderer for this either lol

one more thing in this commit: it seems that there's a weird mechanism in the bind functions that blocks signals from being emitted when loading data? i find the behaviour this creates extremely counterintuitive, as it makes it really annoying to reset checkboxes toggling certain widgets with a regular signal connect, so i've taken the liberty of deleting it for checkboxes for now. for consistency's sake, remind me to remove it for the rest of the binds later and test if everything still works fine. this smells like a workaround for one of this codebase's myriad cursed problems that i'm sure i could fix more properly.