modify max db entry count
hate. let me tell you how much i've come to hate the qt model/view framework since i began to write qt apps. there are over 100 lines of model related code within this commit. if the word "hate" was written on each nanoangstrom of the vector representations of the characters in these lines it would not equal one one-billionth of the hate i feel for the qt model/view framework at this micro-instant. hate. hate.

seriously though. interacting with qmodelindex is a pain in the ass and the whole thing is a black box of mystery segfaults. though maybe the latter also stems from the extremely esoteric Q_OBJECT-macro-doesn't-support-template-classes bypass this model is placed within... not even the light of clangd shines here. i have to blindly swat at the code and find my bugs through compilation errors.

i'm not looking forward to ever touching this code again. unfortunately i'll have to for the filter function. or not. maybe a brave volunteer will dare venture into the belly of the beast for me?

enough distractions. chipsets next.