tileset widget, part 1
i've grown and matured as a person. i will be calling it chipset internally for liblcf parity. you win this time!

i found out while working on this that there is a pr on the base editor for a tileset tab containing some animated tile stuff. unfortunately i'm well past the point where i can retrofit any of its code for this. i've officially given up on making this upstream, i'm too many rewrites in...

speaking of. i'll probably have to offload the model storage to a bespoke model holder for the database and project. the whole point of the model view framework is to have one model in multiple views so things auto-update cleanly. that's another major rewrite that'll affect every database tab though, so i'll be holding off for the sake of getting this commit out as a checkpoint. i also kinda want to show that the fork's still alive. hello, person working on event stuff in a fork of my fork! sorry for putting you in rebase hell for the forseeable future!