tileset widget part 3.5
i love genericizing code! turns out "empty picker that just shows an image" is used for like 7 or more types of images, so this is nice to have. maybe i'll make a fancier one for pictures that lets you preview spritesheet slicing? i dunno. we're sticking with this for now.

also includes a patch to add override for insert/removeRows on the RpgModel. i should really run this thing through with a clang format/tidy sometime

also also. remind me to make the database auto-select the first item in each model. none of these like it when they have no data loaded